+++ pkg/web/pages.go
@@ -84,6 +84,10 @@ func (s *Server) SendEnvelopeForm(c *gin.Context) {
 	c.HTML(http.StatusOK, "send_envelope.html", scene.New(c))
 }
 
+func (s *Server) SendMessageForm(c *gin.Context) {
+	c.HTML(http.StatusOK, "send_message.html", scene.New(c))
+}
+
 func (s *Server) TravelAddressUtility(c *gin.Context) {
 	c.HTML(http.StatusOK, "traveladdress.html", scene.New(c))
 }
