+++ pkg/web/routes.go
@@ -101,6 +101,7 @@ func (s *Server) setupRoutes() (err error) {
 	s.router.GET("/transactions/:id/repair", authenticate, s.TransactionsRepairPreview)
 	s.router.GET("/transactions/:id/info", authenticate, s.TransactionsInfo)
 	s.router.GET("/send-envelope", authenticate, s.SendEnvelopeForm)
+	s.router.GET("/send-message", authenticate, s.SendMessageForm)
 	s.router.GET("/accounts", authenticate, s.Accounts)
 	s.router.GET("/counterparty", authenticate, s.CounterpartyVasps)
 	s.router.GET("/users", authenticate, s.UsersManagement)
