+++ pkg/web/templates/components/beneficiary_information.html
@@ -1,7 +1,7 @@
 {{ define "beneficiary_information" }}
 
 <section class="py-2 border-t border-t-black">
-  <h2 class="my-4 font-bold">2. BENEFICIARY INFORMATION</h2>
+  <h2 class="my-4 font-bold">BENEFICIARY INFORMATION</h2>
   <div class="my-4">
     <label for="travel_address" class="label-style">Travel Address</label>
     <input type="text" id="travel_address" name="travel_address" placeholder="Enter beneficiary's travel address" class="input-style" />
