+++ pkg/web/templates/components/originator_information.html
@@ -1,11 +1,8 @@
 {{ define "originator_information" }}
 
 <section class="py-2 border-t border-t-black">
-  <h2 class="my-4 font-bold">1. ORIGINATOR INFORMATION</h2>
-  <div class="my-5">
-    <label for="orig_crypto_address" class="label-style">Crypto Address</label>
-    <input type="text" id="orig_crypto_address" name="orig_crypto_address" placeholder="Enter originator's crypto address" class="input-style" />
-  </div>
+  <!-- TODO: Add customer lookup to originator section -->
+  <h2 class="my-4 font-bold">ORIGINATOR INFORMATION</h2>
   <div class="grid gap-6 my-4 md:grid-cols-2">
     <div>
       <label for="orig_first_name" class="label-style">Legal First or Given Name</label>
