+++ pkg/web/templates/components/sunrise_beneficiary.html
@@ -1,0 +1,27 @@
+{{ define "sunrise_beneficiary" }}
+
+<section class="py-2 border-t border-t-black">
+  <h2 class="my-4 font-bold">BENEFICIARY INFORMATION</h2>
+  <div class="grid gap-6 my-4 md:grid-cols-2">
+    <div>
+      <label for="benf_first_name" class="label-style">Legal First or Given Name</label>
+      <input type="text" id="benf_first_name" name="benf_first_name" placeholder="Enter beneficiary's first or given name" class="input-style" />
+    </div>
+    <div>
+      <label for="benf_last_name" class="label-style">Legal Last or Family Name</label>
+      <input type="text" id="benf_last_name" name="benf_last_name" placeholder="Enter beneficiary's last or family name" class="input-style" />
+    </div>
+  </div>
+  <div class="grid gap-6 my-4 md:grid-cols-2">
+    <div>
+      <label for="benf_crypto_address" class="label-style">Crypto Address</label>
+      <input type="text" id="benf_crypto_address" name="benf_crypto_address" placeholder="Enter beneficiary's crypto address" class="input-style" />
+    </div>
+    <div>
+      <label for="benf_countries" class="label-style">Country</label>
+      <select id="benf_countries" name="benf_country"></select>
+    </div>
+  </div>
+</section>
+
+{{ end }}
