+++ pkg/web/templates/components/virtual_asset.html
@@ -1,7 +1,7 @@
 {{ define "virtual_asset" }}
 
 <section class="py-2 border-t border-t-black">
-  <h2 class="my-4 font-bold">3. VIRTUAL ASSET TRANSFER DETAILS</h2>
+  <h2 class="my-4 font-bold">VIRTUAL ASSET TRANSFER DETAILS</h2>
   <div class="grid gap-6 my-4 md:grid-cols-2">
     <div>
       <label for="amount" class="label-style">Transfer Amount</label>
