+++ pkg/web/templates/send_envelope.html
@@ -6,7 +6,11 @@
   <section class="my-4 md:m-8">
     <p class="my-2 md:text-lg">Create a secure envelope to share transfer details with the counterparty.</p>
     <form id="secure-envelope-form" hx-post="/v1/transactions/prepare" hx-target="#preview_envelope" hx-indicator="#loader" hx-ext="json-enc" method="post">
-      
+      <div class="my-5">
+        <label for="orig_crypto_address" class="label-style">Crypto Address</label>
+        <input type="text" id="orig_crypto_address" name="orig_crypto_address" placeholder="Enter originator's crypto address" class="input-style" />
+      </div>
+
       {{ template "originator_information" . }}
 
       {{ template "beneficiary_information" . }}
@@ -29,5 +33,6 @@
 {{ end }}
 
 {{ define "appcode" }}
+<script type="module" src="/static/js/msgSelElements.js"></script>
 <script type="module" src="/static/js/envelope.js"></script>
 {{ end }}
